Moved all types to a seperate file
